Extrapolate Orderby and special comparisons from query visitor test
