+++ QueryBuilder_Cassandra/src/test/java/net/sf/esfinge/querybuilder/cassandra/unit/queryvisitor/CassandraOrderByQueryVisitorTest.java
@@ -1,0 +1,89 @@
+package net.sf.esfinge.querybuilder.cassandra.unit.queryvisitor;
+
+import net.sf.esfinge.querybuilder.cassandra.CassandraQueryRepresentation;
+import net.sf.esfinge.querybuilder.cassandra.CassandraVisitorFactory;
+import net.sf.esfinge.querybuilder.cassandra.querybuilding.ordering.OrderByClause;
+import net.sf.esfinge.querybuilder.exception.InvalidQuerySequenceException;
+import net.sf.esfinge.querybuilder.methodparser.ComparisonType;
+import net.sf.esfinge.querybuilder.methodparser.OrderingDirection;
+import net.sf.esfinge.querybuilder.methodparser.QueryRepresentation;
+import net.sf.esfinge.querybuilder.methodparser.QueryVisitor;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class CassandraOrderByQueryVisitorTest {
+
+    private final QueryVisitor visitor = CassandraVisitorFactory.createQueryVisitor();
+
+    @Test
+    public void singleOrderByTest() {
+        visitor.visitEntity("Person");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitEnd();
+
+        OrderByClause expected = new OrderByClause("age", OrderingDirection.ASC);
+
+        QueryRepresentation qr = visitor.getQueryRepresentation();
+        String query = qr.getQuery().toString();
+
+        assertEquals(query, "SELECT * FROM <#keyspace-name#>.Person");
+        assertEquals(expected, ((CassandraQueryRepresentation) qr).getOrderByClause().get(0));
+    }
+
+    @Test
+    public void doubleOrderByTest() {
+        visitor.visitEntity("Person");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitOrderBy("name", OrderingDirection.DESC);
+        visitor.visitEnd();
+
+        OrderByClause expected1 = new OrderByClause("age", OrderingDirection.ASC);
+        OrderByClause expected2 = new OrderByClause("name", OrderingDirection.DESC);
+
+        QueryRepresentation qr = visitor.getQueryRepresentation();
+        String query = qr.getQuery().toString();
+
+        assertEquals(query, "SELECT * FROM <#keyspace-name#>.Person");
+        assertEquals(expected1, ((CassandraQueryRepresentation) qr).getOrderByClause().get(0));
+        assertEquals(expected2, ((CassandraQueryRepresentation) qr).getOrderByClause().get(1));
+    }
+
+    @Test(expected = InvalidQuerySequenceException.class)
+    public void wrongOrderByTest() {
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+    }
+
+    @Test(expected = InvalidQuerySequenceException.class)
+    public void orderByAfterConnectorTest() {
+        visitor.visitEntity("Person");
+        visitor.visitCondition("name", ComparisonType.EQUALS);
+        visitor.visitConector("and");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitEnd();
+    }
+
+    @Test(expected = InvalidQuerySequenceException.class)
+    public void entityAfterOrderByTest() {
+        visitor.visitEntity("Person");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitEntity("Person");
+        visitor.visitEnd();
+    }
+
+    @Test(expected = InvalidQuerySequenceException.class)
+    public void conditionAfterOrderByTest() {
+        visitor.visitEntity("Person");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitCondition("name", ComparisonType.EQUALS);
+        visitor.visitEnd();
+    }
+
+    @Test(expected = InvalidQuerySequenceException.class)
+    public void connectorAfterOrderByTest() {
+        visitor.visitEntity("Person");
+        visitor.visitOrderBy("age", OrderingDirection.ASC);
+        visitor.visitConector("and");
+        visitor.visitEnd();
+    }
+}
